example/requirements.txt
Django==1.7
djangorestframework==2.4.2
wsgiref==0.1.2
